split api hook into seperate file, start to try and get search results back (its not working)
