+++ public/app/core/components/NestedFolderPicker/NestedFolderPicker.tsx
@@ -1,35 +1,21 @@
 import { css } from '@emotion/css';
 import { autoUpdate, flip, useClick, useDismiss, useFloating, useInteractions } from '@floating-ui/react';
-import { createSelector } from '@reduxjs/toolkit';
-import { QueryDefinition, BaseQueryFn } from '@reduxjs/toolkit/dist/query';
-import { QueryActionCreatorResult } from '@reduxjs/toolkit/dist/query/core/buildInitiate';
 import debounce from 'debounce-promise';
-import { RequestOptions } from 'http';
 import React, { useCallback, useEffect, useId, useMemo, useRef, useState } from 'react';
 
 import { GrafanaTheme2 } from '@grafana/data';
 import { config } from '@grafana/runtime';
 import { Alert, Icon, Input, LoadingBar, useStyles2 } from '@grafana/ui';
 import { t } from 'app/core/internationalization';
-import {
-  ListFolderArgs,
-  browseDashboardsAPI,
-  skipToken,
-  useGetFolderQuery,
-} from 'app/features/browse-dashboards/api/browseDashboardsAPI';
-import { PAGE_SIZE } from 'app/features/browse-dashboards/api/services';
+import { skipToken, useGetFolderQuery } from 'app/features/browse-dashboards/api/browseDashboardsAPI';
 import { useBrowseLoadingStatus } from 'app/features/browse-dashboards/state';
-import { getPaginationPlaceholders } from 'app/features/browse-dashboards/state/utils';
-import { DashboardViewItemWithUIItems, DashboardsTreeItem } from 'app/features/browse-dashboards/types';
 import { QueryResponse, getGrafanaSearcher } from 'app/features/search/service';
 import { queryResultToViewItem } from 'app/features/search/service/utils';
 import { DashboardViewItem } from 'app/features/search/types';
-import { RootState } from 'app/store/configureStore';
-import { FolderDTO } from 'app/types';
-import { useDispatch, useSelector } from 'app/types/store';
 
 import { getDOMId, NestedFolderList } from './NestedFolderList';
 import Trigger from './Trigger';
+import { useFolderList } from './api';
 import { useTreeInteractions } from './hooks';
 
 export interface NestedFolderPickerProps {
@@ -65,6 +51,29 @@ async function getSearchResults(searchQuery: string) {
   return { ...queryResponse, items };
 }
 
+// interface FolderPickerOptionsOptions {
+//   isOpen: boolean;
+//   folderOpenState: Record<string, boolean>;
+// }
+
+// interface FolderPickerOptionsReturn {
+//   searchTerm: string;
+//   setSearchTerm: (term: string) => void;
+
+//   folderPickerOptions: Array<DashboardsTreeItem<DashboardViewItemWithUIItems>>;
+// }
+
+// function useFolderPickerOptions({ isOpen, folderOpenState }: FolderPickerOptionsOptions) {
+//   const [search, setSearch] = useState('');
+//   const [isSearching, setIsSearching] = useState(false);
+//   const [searchResults, setSearchResults] = useState<DashboardViewItem[] | null>(null);
+
+//   const handleSearch = useCallback((term: string) => {
+//     setSearch(term);
+//     setIsSearching(true);
+//   }, []);
+// }
+
 export function NestedFolderPicker({
   value,
   invalid,
@@ -84,12 +93,13 @@ export function NestedFolderPicker({
   const [isFetchingSearchResults, setIsFetchingSearchResults] = useState(false);
   const [autoFocusButton, setAutoFocusButton] = useState(false);
   const [overlayOpen, setOverlayOpen] = useState(false);
-  const [folderOpenState, setFolderOpenState] = useState<Record<string, boolean>>({});
+  const [foldersOpenState, setFoldersOpenState] = useState<Record<string, boolean>>({});
   const overlayId = useId();
   const [error] = useState<Error | undefined>(undefined); // TODO: error not populated anymore
   const lastSearchTimestamp = useRef<number>(0);
 
-  const [flatTree, fetchFolderPage] = useFolderList(folderOpenState);
+  const isBrowsing = Boolean(overlayOpen && search && searchResults);
+  const [browseFlatTree, fetchFolderPage] = useFolderList(isBrowsing, foldersOpenState);
 
   useEffect(() => {
     if (!search) {
@@ -98,6 +108,7 @@ export function NestedFolderPicker({
     }
     const timestamp = Date.now();
     setIsFetchingSearchResults(true);
+
     debouncedSearch(search).then((queryResponse) => {
       // Only keep the results if it's was issued after the most recently resolved search.
       // This prevents results showing out of order if first request is slower than later ones.
@@ -143,13 +154,13 @@ export function NestedFolderPicker({
 
   const handleFolderExpand = useCallback(
     async (uid: string, newOpenState: boolean) => {
-      setFolderOpenState((old) => ({ ...old, [uid]: newOpenState }));
+      setFoldersOpenState((old) => ({ ...old, [uid]: newOpenState }));
 
-      if (newOpenState && !folderOpenState[uid]) {
+      if (newOpenState && !foldersOpenState[uid]) {
         fetchFolderPage(uid);
       }
     },
-    [fetchFolderPage, folderOpenState]
+    [fetchFolderPage, foldersOpenState]
   );
 
   const handleFolderSelect = useCallback(
@@ -188,12 +199,32 @@ export function NestedFolderPicker({
 
   // JOSH TODO: return search results as a flat tree
 
+  const flatTree = useMemo(() => {
+    if (isBrowsing) {
+      return browseFlatTree;
+    }
+
+    const searchFlatTree =
+      searchResults?.items.map((item) => ({
+        isOpen: false,
+        level: 0,
+        item: {
+          kind: 'folder' as const,
+          title: item.title,
+          uid: item.uid,
+        },
+      })) ?? [];
+
+    return searchFlatTree;
+  }, [browseFlatTree, isBrowsing, searchResults]);
+
   const isItemLoaded = useCallback(
     (itemIndex: number) => {
       const treeItem = flatTree[itemIndex];
       if (!treeItem) {
         return false;
       }
+
       const item = treeItem.item;
       const result = !(item.kind === 'ui' && item.uiKind === 'pagination-placeholder');
 
@@ -333,180 +364,4 @@ const getStyles = (theme: GrafanaTheme2) => {
       },
     }),
   };
-};
-
-type TODOFolderListPage = ReturnType<typeof listFoldersSelector>;
-type TODORequestPromise = QueryActionCreatorResult<
-  QueryDefinition<ListFolderArgs, BaseQueryFn<RequestOptions>, 'getFolder', FolderDTO[], 'browseDashboardsAPI'>
->;
-
-const createListFoldersSelector = createSelector(
-  [
-    (parentUid: ListFolderArgs['parentUid']) => parentUid,
-    (parentUid: ListFolderArgs['parentUid'], page: ListFolderArgs['page']) => page,
-    (parentUid: ListFolderArgs['parentUid'], page: ListFolderArgs['page'], limit: ListFolderArgs['limit']) => limit,
-  ],
-  (parentUid, page, limit) => {
-    return browseDashboardsAPI.endpoints.listFolders.select({ parentUid, page, limit });
-  }
-);
-
-const listFoldersSelector = createSelector(
-  (state: RootState) => state,
-  (
-    state: RootState,
-    parentUid: ListFolderArgs['parentUid'],
-    page: ListFolderArgs['page'],
-    limit: ListFolderArgs['limit']
-  ) => createListFoldersSelector(parentUid, page, limit),
-  (state, selectFolderList) => selectFolderList(state)
-);
-
-const listAllFoldersSelector = createSelector(
-  [(state: RootState) => state, (state: RootState, requests: TODORequestPromise[]) => requests],
-  (state: RootState, requests: TODORequestPromise[]) => {
-    const seenRequests = new Set<string>();
-
-    const rootPages: TODOFolderListPage[] = [];
-    const pagesByParent: Record<string, TODOFolderListPage[]> = {};
-
-    for (const req of requests) {
-      if (seenRequests.has(req.requestId)) {
-        continue;
-      }
-
-      const page = listFoldersSelector(state, req.arg.parentUid, req.arg.page, req.arg.limit);
-
-      const parentUid = page.originalArgs?.parentUid;
-      if (parentUid) {
-        if (!pagesByParent[parentUid]) {
-          pagesByParent[parentUid] = [];
-        }
-
-        pagesByParent[parentUid].push(page);
-      } else {
-        rootPages.push(page);
-      }
-    }
-
-    return {
-      rootPages,
-      pagesByParent,
-    };
-  }
-);
-
-/**
- * Returns the whether the set of pages are 'fully loaded', and the last page number
- */
-function getPagesLoadStatus(pages: TODOFolderListPage[]): [boolean, number] {
-  const lastPage = pages.at(-1);
-  const fullyLoaded = lastPage && (lastPage.data?.length ?? 0) < (lastPage.originalArgs?.limit ?? 0);
-
-  return [Boolean(fullyLoaded), lastPage?.originalArgs?.page ?? 1];
-}
-
-/**
- * Returns a loaded folder hierarchy as a flat list and a function to load more pages.
- */
-function useFolderList(openFolders: Record<string, boolean>) {
-  const dispatch = useDispatch();
-
-  // Keep a list of all requests so we can
-  //   a) unsubscribe from them when the component is unmounted
-  //   b) use them to select the responses out of the state
-  const requestsRef = useRef<TODORequestPromise[]>([]);
-
-  const state = useSelector((rootState: RootState) => {
-    return listAllFoldersSelector(rootState, requestsRef.current);
-  });
-
-  // Loads the next page of folders for the given parent UID by inspecting the
-  // state to determine what the next page is
-  const requestNextPage = useCallback(
-    (parentUid: string | undefined) => {
-      const pages = parentUid ? state.pagesByParent[parentUid] : state.rootPages;
-      const [fullyLoaded, pageNumber] = getPagesLoadStatus(pages);
-      if (fullyLoaded) {
-        return;
-      }
-
-      const args = { parentUid, page: pageNumber + 1, limit: PAGE_SIZE };
-      const promise = dispatch(browseDashboardsAPI.endpoints.listFolders.initiate(args));
-
-      // It's important that we create a new array so we can correctly memoize with it
-      requestsRef.current = requestsRef.current.concat([promise]);
-    },
-    [state, dispatch]
-  );
-
-  // Unsubscribe from all requests when the component is unmounted
-  useEffect(() => {
-    return () => {
-      for (const req of requestsRef.current) {
-        req.unsubscribe();
-      }
-    };
-  }, []);
-
-  // Convert the individual responses into a flat list of folders, with level indicating
-  // the depth in the hierarchy.
-  // TODO: this will probably go up in the parent component so it can also do search
-  const treeList = useMemo(() => {
-    function createFlatList(
-      parentUid: string | undefined,
-      pages: TODOFolderListPage[],
-      level: number
-    ): Array<DashboardsTreeItem<DashboardViewItemWithUIItems>> {
-      const flatList = pages.flatMap((page) => {
-        const pageItems = page.data ?? [];
-
-        return pageItems.flatMap((item) => {
-          const folderIsOpen = openFolders[item.uid];
-
-          const flatItem: DashboardsTreeItem<DashboardViewItemWithUIItems> = {
-            isOpen: Boolean(folderIsOpen),
-            level: level,
-            item: {
-              kind: 'folder' as const,
-              title: item.title,
-              uid: item.uid,
-            },
-          };
-
-          const childPages = folderIsOpen && state.pagesByParent[item.uid];
-          if (childPages) {
-            const childFlatItems = createFlatList(item.uid, childPages, level + 1);
-            return [flatItem, ...childFlatItems];
-          }
-
-          return flatItem;
-        });
-      });
-
-      const [fullyLoaded] = getPagesLoadStatus(pages);
-      if (!fullyLoaded) {
-        flatList.push(...getPaginationPlaceholders(PAGE_SIZE, parentUid, level));
-      }
-
-      return flatList;
-    }
-
-    const rootFlatTree = createFlatList(undefined, state.rootPages, 1);
-    rootFlatTree.unshift(ROOT_FOLDER_ITEM);
-
-    return rootFlatTree;
-  }, [state, openFolders]);
-
-  return [treeList, requestNextPage] as const;
-}
-
-const ROOT_FOLDER_ITEM = {
-  isOpen: true,
-  level: 0,
-  item: {
-    kind: 'folder' as const,
-    title: 'Dashboards',
-    uid: '',
-  },
 };
